Fix hydrate bug
